requirements.txt
flask
opencv-python
google-generativeai
pillow
python-dotenv
numpy